tests/wpt/css-tests/css-text-3_dev/xhtml1/css3-text-line-break-baspglwj-110.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html lang="en" xmlns="http://www.w3.org/1999/xhtml"><head>
<meta charset="utf-8" />
<title>CSS3 Text, linebreaks: U+A60F VAI QUESTION MARK</title>
<link href="mailto:ishida@w3.org" rel="author" title="Richard Ishida" />
<link href="https://drafts.csswg.org/css-text-3/#line-breaking" rel="help" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="font dom" name="flags" />
<meta content="[Exploratory] The browser will break a line of text after any Unicode character with the BA Other Terminating Punctuation property." name="assert" />
<style type="text/css">
@font-face {
    font-family: 'csstest_ascii';
    src: url('support/csstest-ascii-webfont.woff') format('woff');
    font-weight: normal;
    font-style: normal;
	}
#breakable { font-family: csstest_ascii; font-size: 25px; width: 800px; line-height: 30px; }
</style>
</head>
<body>



<div class="test">
			<div id="breakable">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa꘏bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb</div>
			</div>


<!--Notes:

Box height: <span id='debugresult'></span><script>document.getElementById('debugresult').innerHTML = document.getElementById('breakable').offsetHeight;</script>

-->
<script>
test(function() {
assert_true(document.getElementById('breakable').offsetHeight &gt; 35);
}, " ");
</script>

<div id="log"></div>



</body></html>
